requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
asyncer==0.0.8
attrs==25.1.0
audioop-lts==0.2.1
backoff==2.2.1
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chainlit==2.6.0
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
discord.py==2.5.2
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.14
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
graphics.py==5.0.1.post1
griffe==1.7.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
Lazify==0.4.0
literalai==0.1.201
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mcp==1.10.1
mdurl==0.1.2
monotonic==1.6
multidict==6.2.0
mypy_extensions==1.1.0
narwhals==1.27.1
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.3
openai==1.93.2
openai-agents==0.1.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-alephalpha==0.40.14
opentelemetry-instrumentation-anthropic==0.40.14
opentelemetry-instrumentation-bedrock==0.40.14
opentelemetry-instrumentation-chromadb==0.40.14
opentelemetry-instrumentation-cohere==0.40.14
opentelemetry-instrumentation-crewai==0.40.14
opentelemetry-instrumentation-google-generativeai==0.40.14
opentelemetry-instrumentation-groq==0.40.14
opentelemetry-instrumentation-haystack==0.40.14
opentelemetry-instrumentation-lancedb==0.40.14
opentelemetry-instrumentation-langchain==0.40.14
opentelemetry-instrumentation-llamaindex==0.40.14
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-instrumentation-marqo==0.40.14
opentelemetry-instrumentation-mcp==0.40.14
opentelemetry-instrumentation-milvus==0.40.14
opentelemetry-instrumentation-mistralai==0.40.14
opentelemetry-instrumentation-ollama==0.40.14
opentelemetry-instrumentation-openai==0.40.14
opentelemetry-instrumentation-pinecone==0.40.14
opentelemetry-instrumentation-qdrant==0.40.14
opentelemetry-instrumentation-redis==0.55b1
opentelemetry-instrumentation-replicate==0.40.14
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-sagemaker==0.40.14
opentelemetry-instrumentation-sqlalchemy==0.55b1
opentelemetry-instrumentation-threading==0.55b1
opentelemetry-instrumentation-together==0.40.14
opentelemetry-instrumentation-transformers==0.40.14
opentelemetry-instrumentation-urllib3==0.55b1
opentelemetry-instrumentation-vertexai==0.40.14
opentelemetry-instrumentation-watsonx==0.40.14
opentelemetry-instrumentation-weaviate==0.40.14
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.55b1
packaging==24.2
pandas==2.2.3
password-strength==0.0.3.post2
patsy==1.0.1
phonenumbers==9.0.3
pillow==11.1.0
plotly==6.0.0
plotly-express==0.4.1
posthog==3.25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.10.1
pydantic_core==2.33.1
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-multipart==0.0.18
python-socketio==5.13.0
pytz==2025.1
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==8.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.0
rsa==4.9.1
scipy==1.15.2
seaborn==0.13.2
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sse-starlette==2.4.1
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.47.0
streamlit-lottie==0.0.5
syncer==2.0.3
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traceloop-sdk==0.40.14
twilio==9.6.5
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uptrace==1.34.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==0.20.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.23.0
